mimmo: updating cmake tree for tests
Actually there are no tests available. Need to be provided for each module of the library
strictly working on the module containts.

Old tests are collected in the folder oldtests. Need to be reupped and remanaged as examples.
Will be perfect if an example reports with a c++ code and xml-tui dictionary