iogeneric: new feature reference PID to MimmoGeometry
added to MimmoGeometry the possibility to define a reference PID for the
geometry that is going to be read or written on file. It will apply a unique label to all cells of
the geometry. If the geometry is already pidded, all pre-existent pid will be translated with
respect to the reference PID.