geohandlers: CPREWORK - FVMeshSelection expose internal boundaries
added possibility to expose internal interface patch between selection
and bulk mesh. This is meant as the portion of the selection boundary
which does not belong to the boundary of the bulk. This "internal" patch
is treated as a regular boundary patch
BEWARE- cherry picked commit and modified to solve conflict